Client Organize shit
- Create invoro component
- App component is just for testing purposes
- improved mongo seed
