+++ Invoro.Api/.dockerignore
@@ -8,3 +8,4 @@ docker-compose.yml
 docker-compose.*.yml
 */bin
 */obj
+*/Test*
