+++ Invoro.WebApp/Invoro/invoro/src/lib/components/Invoro/Invoro.component.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import Invoro from './Invoro.component';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Invoro />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
