+++ Invoro.WebApp/Invoro/invoro/src/lib/dataModel/Status.ts
@@ -6,22 +6,4 @@ export enum Status {
     Released = "Released"
 }
 
-export class StatusToDisplayConverter{
-    public static getDisplayName(status:Status){
-        switch (status) {
-            case (Status.NotPlanned):
-              return "Not Planned";
-            case (Status.Planned):
-              return "Planned";
-            case (Status.InProgress):
-              return "In Progress";
-            case (Status.ReadySoon):
-              return "Ready Soon";
-            case (Status.Released):
-              return "Released";
-          }
-          throw new Error(`An unknown status: ${status}`);
-        }
-}
-
 export default Status;
