+++ Invoro.WebApp/Invoro/invoro/src/lib/services/FeaturesApi.service.ts
@@ -12,6 +12,7 @@ export default class FeaturesApi {
     }
 
     let responseData: any[] = await response.json();
-    return responseData.map<Feature>(feature => ({ id: feature.id, name: feature.name, status: feature.status }));
+    
+    return (responseData as Feature[]).map<Feature>(feature => new Feature(feature.id,feature.name,feature.status));
   }
 }
