+++ Invoro.WebApp/Invoro/invoro/src/lib/services/StatusToDisplayConverter.service.ts
@@ -1,0 +1,19 @@
+import Status from "../dataModel/Status";
+
+export default class StatusToDisplayConverter{
+    public static getDisplayName(status:Status){
+        switch (status) {
+            case (Status.NotPlanned):
+              return "Not Planned";
+            case (Status.Planned):
+              return "Planned";
+            case (Status.InProgress):
+              return "In Progress";
+            case (Status.ReadySoon):
+              return "Ready Soon";
+            case (Status.Released):
+              return "Released";
+          }
+          throw new Error(`An unknown status: ${status}`);
+        }
+}
